# Use mold linker in nix env on Linux