--- a/flake.nix
+++ b/flake.nix
@@ -66,6 +66,14 @@
           solhint =
             solhintPkg { inherit (prev) buildNpmPackage fetchFromGitHub; };
         })
+
+        # The mold linker is around 50% faster on Linux than the default linker.
+        # This overlays a mkShell that is configured to use mold on Linux.
+        (final: prev: {
+          mkShell = (prev.lib.optionals prev.stdenv.isLinux) prev.mkShell.override {
+            stdenv = pkgs.stdenvAdapters.useMoldLinker pkgs.clangStdenv;
+          };
+        })
       ];
       pkgs = import nixpkgs { inherit system overlays; };
       crossShell = { config }: