--- a/flake.nix
+++ b/flake.nix
@@ -69,8 +69,8 @@
 
         # The mold linker is around 50% faster on Linux than the default linker.
         # This overlays a mkShell that is configured to use mold on Linux.
-        (final: prev: {
-          mkShell = (prev.lib.optionals prev.stdenv.isLinux) prev.mkShell.override {
+        (final: prev: prev.lib.optionalAttrs prev.stdenv.isLinux {
+          mkShell = prev.mkShell.override {
             stdenv = pkgs.stdenvAdapters.useMoldLinker pkgs.clangStdenv;
           };
         })