Close #2627
Close #1564

This is a follow up from discussion in #2559 

Also adds a workflow to test the nix dev env on the CI, to make sure it's not accidentally broken by making changes to flake.nix.

The change does not affect the dev-env on macOS.

To test the impact the following script can be used

```bash
#!/usr/bin/env bash
set -euTo pipefail

trap '! [[ "$BASH_COMMAND" =~ ^(echo|bench) ]] &&
      printf "+ %s\n" "$BASH_COMMAND"' DEBUG

bench() {
    echo "Ensuring all dependencies are compiled"
    nix develop -c cargo build --locked -q
    nix develop -c cargo nextest run --locked --cargo-quiet --no-run

    echo "Touching all lib.rs files to ensure recompilation"
    find . -path '*/src/lib.rs' -exec touch {} \;

    echo "Timing cargo build on crates in this repo"
    nix develop -c time -p cargo build --locked -q

    echo "Check if mold linker was used:"
    readelf -p .comment target/nix/debug/sequencer | grep mold && echo "mold liker used" || echo "mold linker not used"
    echo

    echo "Timing nextest build"
    nix develop -c time -p cargo nextest run --locked --cargo-quiet --no-run
}

git checkout main
bench

git checkout ma/nix-env-use-mold
bench
```

For me this reduces the compilation time
- for `cargo build` from 51 seconds to 20 seconds 
- for `nextest` build from 64 seconds to 34 seconds.

Full script log:

```
Already on 'main'
Your branch is up to date with 'origin/main'.
Ensuring all dependencies are compiled
+ nix develop -c cargo build --locked -q
+ nix develop -c cargo nextest run --locked --cargo-quiet --no-run
Touching all lib.rs files to ensure recompilation
+ find . -path '*/src/lib.rs' -exec touch {} \;
Timing cargo build on crates in this repo
+ nix develop -c time -p cargo build --locked -q
real 50.74
user 338.16
sys 278.61
Check if mold linker was used:
+ readelf -p .comment target/nix/debug/sequencer
+ grep mold
mold linker not used

Timing nextest build
+ nix develop -c time -p cargo nextest run --locked --cargo-quiet --no-run
real 63.64 
user 469.64
sys 361.95
+ git checkout ma/nix-env-use-mold
Switched to branch 'ma/nix-env-use-mold'
Ensuring all dependencies are compiled
+ nix develop -c cargo build --locked -q
+ nix develop -c cargo nextest run --locked --cargo-quiet --no-run
Touching all lib.rs files to ensure recompilation
+ find . -path '*/src/lib.rs' -exec touch {} \;
Timing cargo build on crates in this repo
+ nix develop -c time -p cargo build --locked -q
real 20.12
user 50.24
sys 26.84
Check if mold linker was used:
+ readelf -p .comment target/nix/debug/sequencer
+ grep mold
  [    47]  mold 2.36.0 (compatible with GNU ld)
mold liker used

Timing nextest build
+ nix develop -c time -p cargo nextest run --locked --cargo-quiet --no-run
real 34.03
user 193.26
sys 89.67
```

cc @tbro @ss-es